Admin now supported by dashboard page
Admin login to dashboard enabled.

Loads page with admin as currentUser, an array of accounts containing userId and Lastname, Firstname values for available users in the system, and a select to choose an account.

When an account is selected, the settings page loads into a div in the middle of the dashboard page.